requirements.txt
numpy==2.2.6
scipy==1.15.3
pandas==2.1.1
matplotlib==3.10.8
opencv-python==4.12.0.88
torch==2.9.1
torchvision==0.24.1
timm==1.0.22
segmentation_models_pytorch==0.5.0
transformers==4.57.3
requests==2.32.5
Pillow==12.0.0
PyYAML==6.0.3
tqdm==4.67.1